`Add new string entry for HighlightingSourceSnapshotLocation`
` `

`Added a new string entry in the Heracles.sln.DotSettings.user file that specifies the location of the HighlightingSourceSnapshotLocation in the user's local system. This change is related to ReSharper's code coverage feature, which collects data about which parts of the code are being covered while running unit tests. No code was removed in this change.`