`Added unique check methods and removed comments`
In `EquipmentRepository.cs`, the `IsTypeUnique` method was modified and its explanatory comment was removed. This method now checks if a given equipment type is unique in the database, returning true if it is and false otherwise.

New methods `IsNameUnique` and `IsUnique` were added to `MuscleGroupRepository.cs` and `WorkoutSessionRepository.cs` respectively. These methods check if a given muscle group name or workout session name is unique in the database, returning true if it is and false otherwise. The comments explaining the previous methods were removed.