Updated User class and refactored middleware and controllers
Updated the `User` class in `User.cs` to remove the requirement for `UserId`, `Name`, and `Email` properties and changed their default values to `null`. Added a parameterless constructor and a constructor with parameters for these properties.

Refactored `UserFromClaims.cs` to use the `User` class from `Heracles.Domain.Users.Models` and changed the namespace to `Heracles.Infrastructure.Helpers`. Updated `GetUserRolesOrDefault` method to return a `List<string>` instead of `List<UserRoles>`.

Renamed `UserInformationMiddleware` class to `ClaimsToUserMiddleware` in `ClaimsToUserMiddleware.cs` and made `ExtractUserId` method static.

Updated `Program.cs` to use `ClaimsToUserMiddleware` instead of `UserInformationMiddleware`.

Added an `Email` property to the `User` object in the `TestUsersController` class in `TestUsersController.cs`.